Add files via upload
What happens to gradient descent and stochastic gradient descent when it estimates a simple linear regression? How many iterations does it take for a sample size of 10 000, 1000, 100?

What happens if the data is generated as ordinary, normal distribution? What happens if it's generated with t(4) errors? All required moments exist, but only just. This is fat-tailed/leptokurtotic data. What happens if the errors are actually skewed? 

See http://rex-analytics.com/gradient-descent-vs-stochastic-gradient-descent-observations-behaviour/ for explanation.